# QUICKSTEP-70-71 Improve aggregation performance