… expressions/aggregation/AggregationHandle.hpp
… expressions/aggregation/AggregationHandleAvg.hpp
… expressions/aggregation/AggregationHandleCount.cpp
… expressions/aggregation/AggregationHandleCount.hpp
… expressions/aggregation/AggregationHandleDistinct.cpp
… expressions/aggregation/AggregationHandleDistinct.hpp
… expressions/aggregation/AggregationHandleMax.cpp
… expressions/aggregation/AggregationHandleMax.hpp
… expressions/aggregation/AggregationHandleMin.cpp
… expressions/aggregation/AggregationHandleMin.hpp
… expressions/aggregation/AggregationHandleSum.cpp
… expressions/aggregation/AggregationHandleSum.hpp
… expressions/aggregation/CMakeLists.txt
… expressions/aggregation/tests/AggregationHandleAvg_unittest.cpp
… expressions/aggregation/tests/AggregationHandleCount_unittest.cpp
… expressions/aggregation/tests/AggregationHandleMax_unittest.cpp
… expressions/aggregation/tests/AggregationHandleMin_unittest.cpp
… expressions/aggregation/tests/AggregationHandleSum_unittest.cpp
… query_optimizer/CMakeLists.txt
… query_optimizer/ExecutionGenerator.cpp
… relational_operators/CMakeLists.txt
… relational_operators/FinalizeAggregationOperator.cpp
… relational_operators/FinalizeAggregationOperator.hpp
… relational_operators/InitializeAggregationOperator.cpp
… relational_operators/InitializeAggregationOperator.hpp
… storage/AggregationOperationState.cpp
… storage/AggregationOperationState.hpp
… storage/CMakeLists.txt
… storage/CollisionFreeVectorTable.cpp
… storage/CollisionFreeVectorTable.hpp