@jianqiao 2017-02-05
Currently there is a gflag for setting the range upbound:
https://github.com/apache/incubator-quickstep/pull/179/files#diff-3d4b5df01ed8edbe255e096eefbfc342R440