@zuyu 2017-02-07
Please resync with the master branch, and I will merge it. Thanks.