@jianqiao 2017-02-07
There is a `CMakeLists` to be updated -- do not merge at this moment.