@pateljm 2017-02-07
Very impressive algorithmic change @jianqiao !!